Added kafka

+++ common-service/crud-service/src/test/java/org/go/together/test/TestUtils.java
@@ -69,7 +69,7 @@ public class TestUtils {
             ManyJoinDto manyJoinDto = new ManyJoinDto();
             manyJoinDto.setId(id);
             manyJoinDto.setName(name);
-            manyJoinDto.setNumber(i);
+            manyJoinDto.setNumber((long) i);
             manyJoinDtos.add(manyJoinDto);
         }
         return manyJoinDtos;
